requirements.txt
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
scipy>=1.10.0
matplotlib>=3.7.0
scikit-image>=0.20.0
reportlab>=4.0.0
python-docx>=1.0.0
python-pptx>=0.6.0
